Fixed NullPointer when calling auth from outside controller
